updated relapse_interval script validation due to error in clinical validation: 'failed to run script validation, check script and the input '. All unit tests pass
